Added Flag Part - Black Triangle with Yellow Border
About the Yellow Triangle:

We start by figuring out where our yellow flag part triangle begins and where the middle of the flag is.

Then, we draw the yellow triangle from the calculated starting point up to the middle. It's something like filling in a triangle with a yellow color.

Black Triangle Inside the Yellow One:

After this, we want to put a black triangle inside the yellow one. So, we calculated where the black triangle should start and end.

After this we recalculate the middle, and then, we print the black triangle within the yellow one.